+++ src/app/demo/components-overview/chip-authentic-example/chip-authentic-example.component.ts
@@ -45,6 +45,7 @@ export class ChipAuthenticExampleComponent {
   groupLabels = ['Group Chip #1', 'Group Chip #2', 'Group Chip #3'];
   labelLongFormat = false;
   denseFormat = false;
+  hideBorder = false;
 
   constructor() {
     this.form = new FormGroup({